Automate generation of final link command.
Final link command is based on two extconf_variables.yaml files,
one for the base c++ wrapper, and one for the specific project.
These two files were being outputted anyway for diagnostic reasons,
so a bit of reformatting allows them to be used as input data
for this process without  impacting other subsystems.